# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniRender Mobile API — real-world city data pipeline
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Building {
  id: string;
  name?: string | null;
  lat: number;
  lon: number;
  height?: number | null;
  levels?: number | null;
  type?: string | null;
}

export interface Road {
  id: string;
  name?: string | null;
  type: string;
  length?: number | null;
}

export interface PointOfInterest {
  id: string;
  name?: string | null;
  category: string;
  lat: number;
  lon: number;
}

export interface CityData {
  city_name: string;
  lat: number;
  lon: number;
  radius: number;
  buildings_count: number;
  roads_count: number;
  pois_count: number;
  buildings: Building[];
  roads: Road[];
  pois: PointOfInterest[];
  unity_ready: boolean;
  fetch_time_ms: number;
}

export interface ElevationData {
  lat: number;
  lon: number;
  elevation_m: number;
}

export interface WeatherData {
  lat: number;
  lon: number;
  temperature_c: number;
  condition: string;
  wind_speed_kmh: number;
  is_day: boolean;
  weather_code: number;
}

export type GetCityDataParams = {
/**
 * Latitude
 */
lat: number;
/**
 * Longitude
 */
lon: number;
/**
 * Search radius in meters
 */
radius?: number;
};

export type GetElevationParams = {
lat: number;
lon: number;
};

export type GetWeatherParams = {
lat: number;
lon: number;
};