# lib/api-zod/src/generated/types/building.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniRender Mobile API — real-world city data pipeline
 * OpenAPI spec version: 0.1.0
 */

export interface Building {
  id: string;
  name?: string | null;
  lat: number;
  lon: number;
  height?: number | null;
  levels?: number | null;
  type?: string | null;
}