# lib/api-zod/src/generated/types/cityData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniRender Mobile API — real-world city data pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { Building } from './building';
import type { PointOfInterest } from './pointOfInterest';
import type { Road } from './road';

export interface CityData {
  city_name: string;
  lat: number;
  lon: number;
  radius: number;
  buildings_count: number;
  roads_count: number;
  pois_count: number;
  buildings: Building[];
  roads: Road[];
  pois: PointOfInterest[];
  unity_ready: boolean;
  fetch_time_ms: number;
}