# lib/api-zod/src/generated/types/road.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniRender Mobile API — real-world city data pipeline
 * OpenAPI spec version: 0.1.0
 */

export interface Road {
  id: string;
  name?: string | null;
  type: string;
  length?: number | null;
}