# lib/api-zod/src/generated/types/weatherData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniRender Mobile API — real-world city data pipeline
 * OpenAPI spec version: 0.1.0
 */

export interface WeatherData {
  lat: number;
  lon: number;
  temperature_c: number;
  condition: string;
  wind_speed_kmh: number;
  is_day: boolean;
  weather_code: number;
}